refactor(goal_planner): remove unused header and divide ThreadSafeData to another file (#8990)
