+++ planning/behavior_path_planner/autoware_behavior_path_goal_planner_module/include/autoware/behavior_path_goal_planner_module/pull_over_planner/freespace_pull_over.hpp
@@ -18,19 +18,14 @@
 #include "autoware/behavior_path_goal_planner_module/pull_over_planner/pull_over_planner_base.hpp"
 
 #include <autoware/freespace_planning_algorithms/abstract_algorithm.hpp>
-#include <autoware/freespace_planning_algorithms/astar_search.hpp>
-#include <autoware/freespace_planning_algorithms/rrtstar.hpp>
 
 #include <tier4_planning_msgs/msg/path_with_lane_id.hpp>
 
 #include <memory>
-#include <vector>
 
 namespace autoware::behavior_path_planner
 {
 using autoware::freespace_planning_algorithms::AbstractPlanningAlgorithm;
-using autoware::freespace_planning_algorithms::AstarSearch;
-using autoware::freespace_planning_algorithms::RRTStar;
 
 class FreespacePullOver : public PullOverPlannerBase
 {
