+++ planning/behavior_path_planner/autoware_behavior_path_goal_planner_module/src/pull_over_planner/geometric_pull_over.cpp
@@ -78,8 +78,6 @@ std::optional<PullOverPath> GeometricPullOver::plan(
   if (!pull_over_path_opt) {
     return {};
   }
-  auto & pull_over_path = pull_over_path_opt.value();
-
-  return pull_over_path;
+  return pull_over_path_opt.value();
 }
 }  // namespace autoware::behavior_path_planner
